@@ -15,13 +15,13 @@
   "author": "Mads Hvelplund <mads@swissarmyronin.dk>",
   "license": "MIT",
   "dependencies": {
-    "commander": "^6.2.0",
-    "debug": "^4.2.0"
+    "commander": "^8.0.0",
+    "debug": "^4.3.2"
   },
   "devDependencies": {
-    "@tsconfig/node14": "^1.0.0",
-    "@types/node": "^14.14.37",
-    "typescript": "^4.2.3"
+    "@tsconfig/node14": "^1.0.1",
+    "@types/node": "^16.3.1",
+    "typescript": "^4.3.5"
   },
   "bin": {
     "terraform-graph2": "./dist/index.js"
